third_party/blink/web_tests/fast/forms/time/time-input-type-expected.txt

Tests for writing and reading .type property of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.type for "time" is correctly "time".
PASS input.type for "TIME" is correctly "time".
PASS input.type for " time " is correctly "text".
PASS successfullyParsed is true

TEST COMPLETE
